catalog

+++ backend/src/main/java/com/example/bootcamp/repository/ShopRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @Repository
 public interface ShopRepository extends JpaRepository<ShopsEntity, Long> {
     Optional<ShopsEntity> findByShopSlug(String shopSlug);
-    Optional<ShopsEntity> findByUserId(Long userId);
+    Optional<ShopsEntity> findByUserId(Integer userId);
     boolean existsByShopName(String shopName);
     boolean existsByShopSlug(String shopSlug);
 }
